URL Fixes for fs_mark

+++ io/disk/fs_mark.py
@@ -13,14 +13,16 @@
 #
 # Copyright (C) 2003-2004 EMC Corporation
 #
-# fs_mark: Benchmark synchronous/async file creation
 #
 # Ported to avocado by Kalpana S Shetty <kalshett@in.ibm.com>
 # Written by Ric Wheeler <ric@emc.com>
 #   http://prdownloads.sourceforge.net/fsmark/fs_mark-3.3.tar.gz
 
-import os
+"""
+fs_mark: Benchmark synchronous/async file creation
+"""
 
+import os
 from avocado import Test
 from avocado import main
 from avocado.utils import archive
@@ -30,7 +32,7 @@ from avocado.utils.partition import Partition
 from avocado.utils.software_manager import SoftwareManager
 
 
-class fs_mark(Test):
+class FSMark(Test):
 
     """
     The fs_mark program is meant to give a low level bashing to file
@@ -44,15 +46,17 @@ class fs_mark(Test):
         """
 
         smm = SoftwareManager()
-        tarball = self.fetch_asset('http://prdownloads.source'
-                                   'forge.net/fsmark/fs_mark-3.3.tar.gz')
-        archive.extract(tarball, self.srcdir)
-        fs_version = os.path.basename(tarball.split('.tar.')[0])
-        self.sourcedir = os.path.join(self.srcdir, fs_version)
+        tarball = self.fetch_asset('https://github.com/josefbacik/fs_mark/'
+                                   'archive/master.zip')
+        archive.extract(tarball, self.teststmpdir)
+        self.sourcedir = os.path.join(self.teststmpdir, 'fs_mark-master')
         os.chdir(self.sourcedir)
         process.run('make')
         build.make(self.sourcedir)
         self.disk = self.params.get('disk', default=None)
+        self.num = self.params.get('num_files', default='1024')
+        self.size = self.params.get('size', default='1000')
+        self.dir = self.params.get('dir', default=self.srcdir)
         self.fstype = self.params.get('fs', default='ext4')
 
         if self.fstype == 'btrfs':
@@ -61,35 +65,36 @@ class fs_mark(Test):
                         smm.install("btrfs-tools"):
                     self.cancel('btrfs-tools is needed for the test to be run')
 
+        if self.disk is not None:
+            self.part_obj = Partition(self.disk, mountpoint=self.dir)
+            self.log.info("Test will run on %s", self.dir)
+            self.log.info("Unmounting the disk before creating file system")
+            self.part_obj.unmount()
+            self.log.info("creating file system")
+            self.part_obj.mkfs(self.fstype)
+            self.log.info("Mounting disk %s on dir %s", self.disk, self.dir)
+            self.part_obj.mount()
+
     def test(self):
         """
         Run fs_mark
         """
         os.chdir(self.sourcedir)
-
-        # Just provide a sample run parameters
-        num_files = self.params.get('num_files', default='1024')
-        size = self.params.get('size', default='1000')
-        self.dir = self.params.get('dir', default=self.teststmpdir)
-
-        self.part_obj = Partition(self.disk, mountpoint=self.dir)
-        self.log.info("Test will run on %s", self.dir)
-        self.log.info("Unmounting the disk/dir before creating file system")
-        self.part_obj.unmount()
-        self.log.info("creating file system")
-        self.part_obj.mkfs(self.fstype)
-        self.log.info("Mounting disk %s on directory %s", self.disk, self.dir)
-        self.part_obj.mount()
-
-        cmd = ('./fs_mark -d %s -s %s -n %s' % (self.dir, size, num_files))
+        cmd = ('./fs_mark -d %s -s %s -n %s' % (self.dir, self.size, self.num))
         process.run(cmd)
 
     def tearDown(self):
         '''
         Cleanup of disk used to perform this test
         '''
-        self.log.info("Unmounting directory %s", self.dir)
-        self.part_obj.unmount()
+        if self.disk is not None:
+            self.log.info("Unmounting disk %s on directory %s",
+                          self.disk, self.dir)
+            self.part_obj.unmount()
+        self.log.info("Removing the filesystem created on %s", self.disk)
+        delete_fs = "dd if=/dev/zero bs=512 count=512 of=%s" % self.disk
+        if process.system(delete_fs, shell=True, ignore_status=True):
+            self.fail("Failed to delete filesystem on %s", self.disk)
 
 
 if __name__ == "__main__":
